put tabs in container to prevent re-render
